Add the rolling reboot tool.
I threw this together today to do a large cluster reboot. There's a lot
more possible with this, such as adding an option to pass a script that
determines host readiness after boot (e.g. don't just wait for the host
to come up, make sure the database has started and is ready too before
moving on to the next host).

Being used in production now.